identify invalid comet designation with fragment
Added code to check to see if comet designation without a fragment is a valid designation or not (e.g., not passing 2024 A-A when 2024 A is not a valid designation).  Basically just checks to see if the remaining designation without the fragment suffix has enough characters to be turned into a normal packed designation.